Refactor: tailwindcss 설정 카테고리별 리팩토링 & 모듈 형태로 내보내기
